requirements.txt
# Browser automation
playwright==1.49.1
playwright-stealth==1.0.6

# Data processing
pandas==2.2.3
numpy==1.26.4

# Visualization
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# Dashboard
streamlit==1.40.2
streamlit-folium==0.23.1
folium==0.18.0

# HTTP & parsing
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# LLM
groq==0.12.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
Pillow==11.0.0

# Report generation
fpdf2==2.8.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
